Temel alt yapı bitti gibi

+++ Business/Handlers/UserClaims/Commands/DeleteUserClaimCommand.cs
@@ -4,6 +4,7 @@ using Business.BusinessAspects;
 using Business.Constants;
 using Core.Aspects.Autofac.Caching;
 using Core.Aspects.Autofac.Logging;
+using Core.CrossCuttingConcerns.Caching;
 using Core.CrossCuttingConcerns.Logging.Serilog.Loggers;
 using Core.Utilities.Results;
 using DataAccess.Abstract;
@@ -19,21 +20,33 @@ namespace Business.Handlers.UserClaims.Commands
         public class DeleteUserClaimCommandHandler : IRequestHandler<DeleteUserClaimCommand, IResult>
         {
             private readonly IUserClaimRepository _userClaimRepository;
+            private readonly ICacheManager _cacheManager;
 
-            public DeleteUserClaimCommandHandler(IUserClaimRepository userClaimRepository)
+            public DeleteUserClaimCommandHandler(IUserClaimRepository userClaimRepository, ICacheManager cacheManager)
             {
                 _userClaimRepository = userClaimRepository;
+                _cacheManager = cacheManager;
             }
 
-            [SecuredOperation(Priority = 1)]
-            [CacheRemoveAspect()]
-            [LogAspect(typeof(FileLogger))]
+            [SecuredOperation("Admin")]
+            [CacheRemoveAspect("GetOrdersQuery")]
+            [LogAspect(typeof(MsSqlLogger))]
             public async Task<IResult> Handle(DeleteUserClaimCommand request, CancellationToken cancellationToken)
             {
                 var entityToDelete = await _userClaimRepository.GetAsync(x => x.UserId == request.Id);
+                
+                if (entityToDelete == null)
+                {
+                    return new ErrorResult("Kullanıcı claim'i bulunamadı.");
+                }
+
+                var userId = entityToDelete.UserId; // UserId'yi kaydet
 
                 _userClaimRepository.Delete(entityToDelete);
                 await _userClaimRepository.SaveChangesAsync();
+
+                // Cache'i temizle - Yetki değişikliği yapıldığı için cache'i güncellemek gerekiyor
+                _cacheManager.Remove($"{CacheKeys.UserIdForClaim}={userId}");
 
                 return new SuccessResult(Messages.Deleted);
             }
